+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Head
+
+* Fixed: again removed `stylelint.utils.ruleTester`.
+
 # 4.3.2
 
 * Fixed: move `tape` to dependencies to support `testUtils`.
@@ -9,6 +13,7 @@
 # 4.3.0
 
 * Added: `font-family-name-quotes` rule.
+* Added: `codeFilename` option to Node API.
 * Added: `font-weight-notation` rule.
 * Added: `media-feature-no-missing-punctuation` rule.
 * Added: `no-duplicate-selectors` rule.